Last commit

+++ src/main/java/com/raf/users/dto/ReservationDto.java
@@ -3,7 +3,7 @@ package com.raf.users.dto;
 public class ReservationDto {
 
     private Long id;
-    private Long numberOfReservations;
+    private int numberOfReservations;
 
     public Long getId() {
         return id;
@@ -13,11 +13,11 @@ public class ReservationDto {
         this.id = id;
     }
 
-    public Long getNumberOfReservations() {
+    public int getNumberOfReservations() {
         return numberOfReservations;
     }
 
-    public void setNumberOfReservations(Long numberOfReservations) {
+    public void setNumberOfReservations(int numberOfReservations) {
         this.numberOfReservations = numberOfReservations;
     }
 }
